move xorm logger bridge from log to models so that log module could be a standalone package
