+++ models/log.go
@@ -2,11 +2,12 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package log
+package models
 
 import (
 	"fmt"
 
+	"code.gitea.io/log"
 	"github.com/go-xorm/core"
 )
 
@@ -16,78 +17,73 @@ type XORMLogBridge struct {
 	level   core.LogLevel
 }
 
-var (
-	// XORMLogger the logger for xorm
-	XORMLogger *XORMLogBridge
-)
-
-// InitXORMLogger inits a log bridge for xorm
-func InitXORMLogger(showSQL bool) {
-	XORMLogger = &XORMLogBridge{
+// NeXORMLogger inits a log bridge for xorm
+func NeXORMLogger(showSQL bool) core.ILogger {
+	return &XORMLogBridge{
 		showSQL: showSQL,
 	}
 }
 
 // GetGiteaLevel returns the minimum Gitea logger level
-func (l *XORMLogBridge) GetGiteaLevel() Level {
-	return GetLogger("xorm").GetLevel()
+func (l *XORMLogBridge) GetGiteaLevel() log.Level {
+	return log.GetLogger("xorm").GetLevel()
 }
 
 // Log a message with defined skip and at logging level
-func (l *XORMLogBridge) Log(skip int, level Level, format string, v ...interface{}) error {
-	return GetLogger("xorm").Log(skip+1, level, format, v...)
+func (l *XORMLogBridge) Log(skip int, level log.Level, format string, v ...interface{}) error {
+	return log.GetLogger("xorm").Log(skip+1, level, format, v...)
 }
 
 // Debug show debug log
 func (l *XORMLogBridge) Debug(v ...interface{}) {
-	l.Log(2, DEBUG, fmt.Sprint(v...))
+	l.Log(2, log.DEBUG, fmt.Sprint(v...))
 }
 
 // Debugf show debug log
 func (l *XORMLogBridge) Debugf(format string, v ...interface{}) {
-	l.Log(2, DEBUG, format, v...)
+	l.Log(2, log.DEBUG, format, v...)
 }
 
 // Error show error log
 func (l *XORMLogBridge) Error(v ...interface{}) {
-	l.Log(2, ERROR, fmt.Sprint(v...))
+	l.Log(2, log.ERROR, fmt.Sprint(v...))
 }
 
 // Errorf show error log
 func (l *XORMLogBridge) Errorf(format string, v ...interface{}) {
-	l.Log(2, ERROR, format, v...)
+	l.Log(2, log.ERROR, format, v...)
 }
 
 // Info show information level log
 func (l *XORMLogBridge) Info(v ...interface{}) {
-	l.Log(2, INFO, fmt.Sprint(v...))
+	l.Log(2, log.INFO, fmt.Sprint(v...))
 }
 
 // Infof show information level log
 func (l *XORMLogBridge) Infof(format string, v ...interface{}) {
-	l.Log(2, INFO, format, v...)
+	l.Log(2, log.INFO, format, v...)
 }
 
 // Warn show warning log
 func (l *XORMLogBridge) Warn(v ...interface{}) {
-	l.Log(2, WARN, fmt.Sprint(v...))
+	l.Log(2, log.WARN, fmt.Sprint(v...))
 }
 
 // Warnf show warnning log
 func (l *XORMLogBridge) Warnf(format string, v ...interface{}) {
-	l.Log(2, WARN, format, v...)
+	l.Log(2, log.WARN, format, v...)
 }
 
 // Level get logger level
 func (l *XORMLogBridge) Level() core.LogLevel {
 	switch l.GetGiteaLevel() {
-	case TRACE, DEBUG:
+	case log.TRACE, log.DEBUG:
 		return core.LOG_DEBUG
-	case INFO:
+	case log.INFO:
 		return core.LOG_INFO
-	case WARN:
+	case log.WARN:
 		return core.LOG_WARNING
-	case ERROR, CRITICAL:
+	case log.ERROR, log.CRITICAL:
 		return core.LOG_ERR
 	}
 	return core.LOG_OFF
